Refactoring Board to abstract class instead of interface.

+++ src/test/java/org/solitaire/board/MoveTest.java
@@ -12,14 +12,14 @@ public class MoveTest {
 
     @Test
     public void testMoveAttributes() {
-        Integer start = 6;
-        Integer center = 5;
-        Integer end = 4;
+        int start = 6;
+        int center = 5;
+        int end = 4;
         Move move = new Move(start, center, end);
-        Long mask = (1L << start) | (1L << center) | (1L << end);
-        Long check = (1L << start) | (1L << center);
-        assertEquals(mask, (Long) move.getMask());
-        assertEquals(check, (Long) move.getCheck());
+        long mask = (1L << start) | (1L << center) | (1L << end);
+        long check = (1L << start) | (1L << center);
+        assertEquals(mask, move.getMask());
+        assertEquals(check, move.getCheck());
         assertEquals(1L << start, move.getStart());
         assertEquals(1L << end, move.getEnd());
         assertEquals(start, move.getStartID());
